+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.1.2] - Unreleased
+### Added
+- New component `ola-divider`
+- New component `ola-counter`
+
 ## [0.1.1] - 2020-05-08
 ### Removed
 - mjmlconfig file
@@ -16,4 +21,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## 0.1.0 - 2020-04-23
 First version
 
+[0.1.2]: https://github.com/marketgoo/Ola-Emails/compare/v0.1.1...HEAD
 [0.1.1]: https://github.com/marketgoo/Ola-Emails/compare/v0.1.0...v0.1.1